32-bit: narrow, cast, or refuse -- thirteen sites, three different right answers
The `netsim` root cause first: `Prng.below(n: usize) usize` was being
reused for `Time`-typed draws. `Time` is the simulated clock and is
genuinely `u64` on every host, so the answer is `belowWide(n: u64) u64`
-- bound and draw both stay wide, nothing to truncate -- with the
index-typed call sites left on `below` where `usize` is correct. Its
test pins a bound past `maxInt(u32)`, which is the only part of this
that a 64-bit host can actually exercise. All four dependents
(`df-elect`, `liveness-hyst`, `raft`) clear through it.

⚠ `minisign` is the one that had to REFUSE rather than convert.
`raw.mem_limit` comes out of the key file, so it is attacker-controlled
and can genuinely exceed `usize` on a 32-bit host. An `@intCast` there
would have silently downgraded scrypt's memory-hardness -- a security
property quietly weakened by a portability fix, which is about the worst
shape this class can take. It now returns
`OpenSecretKeyError.MemLimitTooLarge`.

The rest divide by what the value actually is, not by what compiles:

  - narrowed to `usize` where the value sizes an allocation or indexes
    memory and the wide type was accidental: `bbs`'s generator index,
    `lockfree`'s `per_producer`, `loopix`'s test counters.
  - cast where the value is provably bounded at the site -- a modulo, a
    scenario-authored length, a small test literal: `isis-sim`,
    `loopfree-reconv` (two moduli; the third restructured to compare
    wide and narrow only after the existing bounds check), `rescue`,
    `yaml`'s fuzz bounds, `zipstream`. Production fields that model real
    file offsets (`zipstream`'s zip64 sizes) stay `u64`.
  - `@alignCast` for `@fieldParentPtr` in `aaa-gate` and `ratelimit`:
    `Entry`'s `u64` fields give it stricter alignment than the intrusive
    `DoublyLinkedList.Node` needs on 32-bit. Same idiom already used in
    `http`, `ipcbus` and `zipstream`, and sound because every `Entry`
    comes from `gpa.create`.

Seven of these modules now pass `portable-*` outright. `bbs`, `lockfree`
and `yaml` still fail it -- on wasi-surface gaps and the separate
64-bit-atomic class, not on anything here. Said plainly rather than
counted as done.

+++ modules/bbs/src/ciphersuite.zig
@@ -221,7 +221,15 @@ pub fn createGeneratorsWithSeed(allocator: std.mem.Allocator, count: usize, seed
     const out = try allocator.alloc(G1.Affine, count);
     errdefer allocator.free(out);
     var v = expandMessage(expand_len, seed_message, seed_dst);
-    var i: u64 = 1;
+    // `i` only ever ranges over `1..=count`, and `count` is already `usize`
+    // (it bounds the `allocator.alloc` above) — there is no value `i` can
+    // hold that a `usize` can't. The `u64` here was accidental width, not a
+    // requirement: the draft's `I2OSP(i, 8)` is the WIRE format (an 8-byte
+    // big-endian field), not a claim that `i` itself needs 64-bit range, and
+    // `std.mem.writeInt(u64, ...)` accepts a `usize` argument via ordinary
+    // implicit widening on every host. Narrowed rather than cast at the
+    // boundary; identical semantics, so no new test.
+    var i: usize = 1;
     while (i <= count) : (i += 1) {
         var input: [expand_len + 8]u8 = undefined;
         input[0..expand_len].* = v;
